refactor: restructure hops-react internals
Split mixins into separate logical groups and colocate their React
components.
