+++ packages/react/import-component/context.js
@@ -1,0 +1,3 @@
+const React = require('react');
+
+module.exports = React.createContext({});
