+++ packages/react/import-component/mixin.core.js
@@ -1,0 +1,12 @@
+const { Mixin } = require('hops-bootstrap');
+
+class ImportComponentCoreMixin extends Mixin {
+  configureBuild(webpackConfig, { jsLoaderConfig }) {
+    jsLoaderConfig.options.plugins.push([
+      require.resolve('../lib/babel'),
+      { module: 'hops' },
+    ]);
+  }
+}
+
+module.exports = ImportComponentCoreMixin;
